+++ src/pages/JoinPage/EmailPwPage.style.js
@@ -29,6 +29,15 @@ export const Input = styled.input`
   font-size: 14px;
   font-weight: 400;
   ${CommonFont}
+  &:-webkit-autofill,
+  &:-webkit-autofill:hover,
+  &:-webkit-autofill:focus,
+  &:-webkit-autofill:active {
+    transition: background-color 5000s ease-in-out 0s;
+    -webkit-transition: background-color 9999s ease-out;
+    -webkit-box-shadow: 0 0 0px 1000px 'var(--white);' inset !important;
+    -webkit-text-fill-color: 'var(--white);' !important;
+  }
 
   &::placeholder {
     color: var(--gray-02);
