+++ src/pages/JoinPage/JoinPage.jsx
@@ -1,16 +1,17 @@
 import React, { useState, useEffect } from 'react';
-import { useLocation, useNavigate } from 'react-router-dom';
-import Header from 'components/publicRoute/Header';
+import EmailPwPage from './EmailPwPage';
+import Header from 'components/login/Header';
+import { Container } from 'pages/JoinPage/EmailPwPage.style';
+import ProfileSettingPage from './ProfileSettingPage';
 import Button from 'components/common/button/Button';
-import { Container, Label, Input, Line, ErrMsg } from 'components/publicRoute/PublicRoutePage.style';
 import { useValidation } from 'hook/useValidation';
 
 export default function JoinPage() {
-  const navigate = useNavigate();
   const { email, password, setEmail, setPassword, loginError, validateLogin, errMsgVisible, setErrMsgVisible } = useValidation();
-  const [disabled, setDisabled] = useState(false);
+  const [emailPwPageVisible, setEmailPwPageVisible] = useState(true);
   const [emailFocus, setEmailFocus] = useState(false);
   const [pwFocus, setPwFocus] = useState(false);
+  const [disabled, setDisabled] = useState(false);
 
   const handleData = (event) => {
     if (event.target.type === 'email') {
@@ -41,53 +42,53 @@ export default function JoinPage() {
     setErrMsgVisible(false);
   };
 
+  const handleSubmit = (event) => {
+    event.preventDefault();
+    if (emailPwPageVisible) {
+      // EmailPwPage가 보이는 상태에서 버튼을 클릭한 경우
+      if (email && password) {
+        // 이메일 및 비밀번호 정보가 입력되었을 때
+        validateLogin();
+        setEmailPwPageVisible(false); // EmailPwPage를 숨기고
+      } else {
+        // 이메일 및 비밀번호 정보가 입력되지 않았을 때
+        setErrMsgVisible(true);
+      }
+    } else {
+      // ProfileSettingPage가 보이는 상태에서 버튼을 클릭한 경우
+      // 이곳에서 추가적인 로직 수행 가능
+    }
+  };
+
   useEffect(() => {
     setDisabled(!(email && password));
   }, [email, password, loginError]);
 
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    validateLogin();
-    localStorage.setItem('email', email);
-    localStorage.setItem('password', password);
-    navigate('/profilesetting');
-  };
-
   return (
     <>
-      <Header text="이메일로 회원가입" isRender={true} />
-      <Container>
-        <form onSubmit={handleSubmit}>
-          <Label htmlFor="user-email">이메일</Label>
-          <Input
-            type="email"
-            id="user-email"
-            value={email}
-            placeholder="이메일 주소를 입력해 주세요."
-            required
-            onChange={handleData}
-            onBlur={handleBlur}
-            onFocus={handleFocus}
-          ></Input>
-          <Line $emailFocus={emailFocus}></Line>
-
-          <Label htmlFor="user-pw">비밀번호</Label>
-          <Input
-            type="password"
-            id="user-pw"
-            value={password}
-            placeholder="비밀번호를 입력해 주세요."
-            autoComplete="current-password"
-            required
-            onChange={handleData}
-            onBlur={handleBlur}
-            onFocus={handleFocus}
-          ></Input>
-          <Line $pwFocus={pwFocus}></Line>
-        </form>
-        <Button size="lg" vari="basic" text="다음" type="submit" onClick={handleSubmit} disabled={disabled} />
-        {errMsgVisible && loginError && <ErrMsg>{loginError}</ErrMsg>}
-      </Container>
+      {emailPwPageVisible && <Header text="이메일로 회원가입" />}
+      {emailPwPageVisible && (
+        <Container>
+          <EmailPwPage
+            email={email}
+            password={password}
+            emailFocus={emailFocus}
+            pwFocus={pwFocus}
+            setEmail={setEmail}
+            setPassword={setPassword}
+            loginError={loginError}
+            validateLogin={validateLogin}
+            errMsgVisible={errMsgVisible}
+            setErrMsgVisible={setErrMsgVisible}
+            handleData={handleData}
+            handleBlur={handleBlur}
+            handleFocus={handleFocus}
+            handleSubmit={handleSubmit}
+          />
+          <Button size="lg" vari="basic" text={'다음'} type="submit" onClick={handleSubmit} disabled={disabled} />
+        </Container>
+      )}
+      {!emailPwPageVisible && <ProfileSettingPage />}
     </>
   );
 }
