+++ src/pages/JoinPage/ProfileSettingPage.style.js
@@ -1,0 +1,34 @@
+import { styled } from 'styled-components';
+
+export const Container = styled.div`
+  width: 100%;
+  border-bottom: none;
+  background-color: var(--white);
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  padding-top: 36px;
+`;
+
+export const Wrap = styled.div`
+  display: flex;
+  flex-direction: column;
+`;
+
+export const ProfileSettingCenterText = styled.h1`
+  margin: 0 auto;
+  margin-bottom: 12px;
+  color: var(--black);
+  font-weight: var(--medium);
+  font-size: 20px;
+`;
+
+export const CenterSubText = styled.p`
+  margin-bottom: 30px;
+  color: var(--gray-01);
+  text-align: center;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 14px; /* 100% */
+`;
